requirements.txt
fastapi>=0.109.0
python-multipart
uvicorn[standard]
numpy<2.0.0
opencv-python
pillow
realesrgan
basicsr>=1.4.2
facexlib
gfpgan
transformers
timm
accelerate
einops
scikit-learn
mediapipe
torch
torchvision
torchaudio
requests
scikit-image
simple-lama-inpainting
supabase
slowapi